+++ packages/expo-screen-capture/android/src/main/java/expo/modules/screencapture/ScreenShotEventEmitter.kt
@@ -6,32 +6,19 @@ import android.content.pm.PackageManager
 import android.database.ContentObserver
 import android.net.Uri
 import android.os.Build
-import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import android.provider.MediaStore
 import android.util.Log
-
-import androidx.core.content.ContextCompat
 import androidx.annotation.Nullable
-
-import expo.modules.core.ModuleRegistry
+import androidx.core.content.ContextCompat
 import expo.modules.core.interfaces.LifecycleEventListener
-import expo.modules.core.interfaces.services.EventEmitter
-import expo.modules.core.interfaces.services.UIManager
 
-import java.lang.Exception
-
-class ScreenshotEventEmitter(val context: Context, moduleRegistry: ModuleRegistry) : LifecycleEventListener {
-  private val onScreenshotEventName: String = "onScreenshot"
+class ScreenshotEventEmitter(val context: Context, onCapture: () -> Unit) : LifecycleEventListener {
   private var isListening: Boolean = true
-  private var eventEmitter: EventEmitter
   private var previousPath: String = ""
 
   init {
-    moduleRegistry.getModule(UIManager::class.java).registerLifecycleEventListener(this)
-    eventEmitter = moduleRegistry.getModule(EventEmitter::class.java)
-
     val contentObserver: ContentObserver = object : ContentObserver(Handler(Looper.getMainLooper())) {
       override fun onChange(selfChange: Boolean, uri: Uri?) {
         super.onChange(selfChange, uri)
@@ -43,7 +30,7 @@ class ScreenshotEventEmitter(val context: Context, moduleRegistry: ModuleRegistr
           val path = getFilePathFromContentResolver(context, uri)
           if (path != null && isPathOfNewScreenshot(path)) {
             previousPath = path
-            eventEmitter.emit(onScreenshotEventName, Bundle())
+            onCapture()
           }
         }
       }